requirements.txt
# 深度学习框架
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
trl>=0.7.0
datasets>=2.14.0

# 科学计算
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.17.0
pandas>=2.1.0

# GPU/DCU支持
torch-rocm

# 大模型相关
tokenizers>=0.14.0
sentencepiece>=0.1.99
protobuf>=4.24.0
bitsandbytes>=0.41.0

# 配置和工具
pyyaml>=6.0.1
toml>=0.10.2
click>=8.1.7
tqdm>=4.66.0
rich>=13.6.0

# 监控和日志
tensorboard>=2.14.0
wandb>=0.15.0
psutil>=5.9.0
gpustat>=1.1.1

# Web和API
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=3.50.0
streamlit>=1.28.0

# 开发工具
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0

# 容器化
docker>=6.1.0
docker-compose>=1.29.0 